+++ s4e-backend/src/main/java/pl/cyfronet/s4e/admin/sync/AdminSceneSyncRequest.java
@@ -1,0 +1,13 @@
+package pl.cyfronet.s4e.admin.sync;
+
+import lombok.Builder;
+import lombok.Data;
+
+import javax.validation.constraints.NotEmpty;
+
+@Data
+@Builder
+public class AdminSceneSyncRequest {
+    @NotEmpty
+    private String prefix;
+}